requirements.txt
python-telegram-bot==20.7
notion-client==2.2.1
APScheduler==3.10.4
pytz==2024.1
aiohttp==3.9.3
